requirements.txt
streamlit
datetime
numpy
pandas
wordcloud
Sastrawi
matplotlib
pillow